pyproject.toml
[tool.poetry]
name = "talkwave"
version = "0.0.5"
description = "TalkWave is an AI chatbot for developers written in Python."
authors = ["Sebastien Rousseau <sebastian.rousseau at gmail.com>"]
license = "Apache Software License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.2.3"
openai = "^0.27.2"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
tabulate = "^0.9.0"

[tool.poetry.dev-dependencies]
coverage="7.2.3"
pytest-cov="4.0.0"
pytest="7.3.0"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
